+++ client/lib/ft4/authentication/evm/key-handler.ts
@@ -4,8 +4,11 @@ import { BufferId } from "../../../cryptoUtils";
 import { hasAuthDescriptorFlags } from "../ft/key-handler";
 import { AuthDataService, KeyHandler, KeyStore } from "../types";
 import { AnyAuthDescriptorRegistration } from "/ft4/accounts/auth-descriptor/types";
-import { TxBuilderTransaction } from "/ft4/utils/types";
 import { gtv, deriveAccountId } from "/ft4/accounts/auth-descriptor";
+import { TxContext, TxBuilderTransaction } from "/ft4/utils/types";
+
+const getNonceId = (accountId: BufferId, authDescriptorId: BufferId) =>
+  accountId.toString("hex") + authDescriptorId.toString("hex");
 
 export function createEvmKeyHandler(
   authDescriptorRegistration: AnyAuthDescriptorRegistration,
@@ -19,7 +22,7 @@ export function createEvmKeyHandler(
     authorize: (
       accountId: BufferId,
       operation: Operation,
-      nonce: number,
+      context: TxContext,
       authDataService: AuthDataService,
     ) =>
       authorize(
@@ -28,8 +31,8 @@ export function createEvmKeyHandler(
           gtv.authDescriptorRegistrationToGtv(authDescriptorRegistration),
         ),
         operation,
-        nonce,
         authDataService,
+        context,
         keyStore,
       ),
     sign: (transaction: TxBuilderTransaction) => sign(transaction, keyStore),
@@ -41,12 +44,20 @@ async function authorize(
   accountId: BufferId,
   authDescriptorId: BufferId,
   operation: Operation,
-  nonce: number,
   authDataService: AuthDataService,
+  context: TxContext,
   keyStore: EvmKeyStore,
 ): Promise<Operation[]> {
-  const messageTemplate =
-    await authDataService.getAuthMessageTemplate(operation);
+  const messageTemplate = await authDataService.getAuthMessageTemplate(
+    operation,
+  );
+  const nonce = await getNonce(
+    authDataService,
+    accountId,
+    authDescriptorId,
+    context,
+  );
+
   const brid = authDataService.getBrid();
   const message = messageTemplate
     .replace("{account_id}", formatter.ensureBuffer(accountId).toString("hex"))
@@ -65,7 +76,34 @@ async function authorize(
 async function sign(
   transaction: TxBuilderTransaction,
   keyStore: KeyStore,
-): Promise<void> {
-  // return transaction.sign(keyStore);
-}
+): Promise<void> {}
 /* eslint-enable */
+
+async function getNonce(
+  authDataService: AuthDataService,
+  accountId: BufferId,
+  authDescriptorId: BufferId,
+  context: TxContext,
+) {
+  let evmContext = context["evm"];
+  if (!evmContext) {
+    evmContext = {
+      nonce: {},
+    };
+
+    context["evm"] = evmContext;
+  } else if (!evmContext.nonce) {
+    evmContext["nonce"] = {};
+  }
+
+  const nonceId = getNonceId(accountId, authDescriptorId);
+  const cachedNonce = evmContext.nonce[nonceId];
+  if (cachedNonce !== 0 && !cachedNonce) {
+    const nonce = await authDataService.getNonce(accountId, authDescriptorId);
+    evmContext.nonce[nonceId] = nonce;
+  } else {
+    evmContext.nonce[nonceId] += 1;
+  }
+
+  return evmContext.nonce[nonceId];
+}